src/app/q8/q8.component.ts
import {
  Component,
  Directive,
  ElementRef,
  HostListener,
  Input,
} from '@angular/core';

@Component({
  selector: 'app-q8',
  templateUrl: './q8.component.html',
  styleUrls: ['./q8.component.scss'],
})
export class Q8Component {}

@Directive({
  selector: '[highlight]',
})
export class Q8HighlightDirective {
  @Input('color') color!: 'red' | 'green' | 'yellow';

  constructor(private el: ElementRef) {}

  @HostListener('mouseenter') onMouseEnter() {
    this.highlight(this.color);
  }

  @HostListener('mouseleave') onMouseLeave() {
    this.highlight('');
  }
  private highlight(color: string) {
    this.el.nativeElement.style.backgroundColor = color;
  }
}
